refactor: make Buyer have getResult method
