+++ src/test/java/buyer/BuyerResultTest.java
@@ -29,11 +29,11 @@ public class BuyerResultTest {
     @Test
     void profitTest() {
         //given
-        List<Lotto> lotteries = new ArrayList<>();
-        lotteries.add(new Lotto(List.of(1, 2, 3, 4, 5, 6)));
+        Buyer buyer = new Buyer(1000);
+        buyer.buyLottery(1000, new Lotto("1, 2, 3, 4, 5, 6"));
         WinningLotto winningLotto = new WinningLotto(List.of(1, 2, 3, 4, 5, 6), 8);
         //when
-        BuyerResult buyerResult = winningLotto.getResult(lotteries);
+        BuyerResult buyerResult = buyer.getResult(winningLotto);
         //then
         Assertions.assertThat(buyerResult.getProfit()).isEqualTo((double) Rank.FIRST.prize / 1000);
