+++ src/test/java/lotto/WinningLottoTest.java
@@ -1,14 +1,10 @@
 package lotto;
 
-import buyer.BuyerResult;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
-import java.util.EnumMap;
 import java.util.List;
-import java.util.Map;
 
 public class WinningLottoTest {
     @DisplayName("당첨 번호 입력 테스트")
@@ -39,35 +35,5 @@ public class WinningLottoTest {
         //when
         //then
         Assertions.assertThat(winningLotto.getRank(lotto)).isEqualTo(Rank.FIFTH);
-    }
-
-    @DisplayName("로또 하나에 대한 결과 객체 확인")
-    @Test
-    void lotteryResult() {
-        //given
-        WinningLotto winningLotto = new WinningLotto(List.of(1, 2, 3, 4, 5, 6), 7);
-        EnumMap<Rank, Integer> cpMap = new EnumMap<>(Map.of(Rank.FIFTH, 1));
-        List<Lotto> lotteries = new ArrayList<>();
-        //when
-        lotteries.add(new Lotto(List.of(1, 2, 3, 12, 13, 14)));
-        BuyerResult buyerResult = winningLotto.getResult(lotteries);
-        //then
-        Assertions.assertThat(buyerResult).isEqualTo(new BuyerResult(cpMap));
-    }
-
-    @DisplayName("로또 뭉치에 대한 결과 객체 확인")
-    @Test
-    void lotteryListResult() {
-        //given
-        WinningLotto winningLotto = new WinningLotto(List.of(1, 2, 3, 4, 5, 6), 7);
-        List<Lotto> lotteries = new ArrayList<>();
-        EnumMap<Rank, Integer> cpMap =
-                new EnumMap<>(Map.ofEntries(Map.entry(Rank.FIFTH, 1), Map.entry(Rank.FOURTH, 1)));
-        //when
-        lotteries.add(new Lotto(List.of(1, 2, 3, 12, 13, 14)));
-        lotteries.add(new Lotto(List.of(1, 2, 3, 4, 12, 14)));
-        BuyerResult buyerResult = winningLotto.getResult(lotteries);
-        //then
-        Assertions.assertThat(buyerResult).isEqualTo(new BuyerResult(cpMap));
     }
 }
